Moved LocalControl/Loader to plugins
