+++ frictionless/loaders/__init__.py
@@ -1,4 +1,3 @@
-from .local import LocalLoader
 from .remote import RemoteLoader
 from .stream import StreamLoader
 from .text import TextLoader
